+++ src/cli/printers/ser/mod.rs
@@ -1,0 +1,4 @@
+mod info;
+mod song_list;
+
+pub use self::{info::*, song_list::*};