+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## future
 ### New features
 - Tab completion now can suggest aliases.
+- Set the print style with `--print`. Add new print styles: `debug`.
 
 ### Fixes
 - Fix image scaling for images that are taller tan wide.